chore: Refractor
