+++ src/krisi/evaluate/score.py
@@ -1,6 +1,6 @@
 from typing import Any, Dict, List, Optional, Union
 
-from krisi.evaluate.library.benchmarking import Model
+from krisi.evaluate.library.benchmarking_models import Model
 from krisi.evaluate.metric import Metric
 from krisi.evaluate.scorecard import ScoreCard
 from krisi.evaluate.type import (
